Add 39 label detachment

+++ app/Services/OrderPaymentLabelsService.php
@@ -80,7 +80,10 @@ readonly class OrderPaymentLabelsService
 
         $sumOfGrossValues = round($totalProductPrice + $additional_service + $additional_cod_cost + $shipment_price_client);
 
-        if (round($this->orderDepositPaidCalculator->calculateDepositPaidOrderData($order)['balance']) !== $sumOfGrossValues) {
+        if (
+            round($this->orderDepositPaidCalculator->calculateDepositPaidOrderData($order)['balance']) !== $sumOfGrossValues &&
+            $order->payments->count() > 0
+        ) {
             RemoveLabelService::removeLabels($order, [39], $arr, [], Auth::user()?->id);
         }
     }
